+++ projects/demo/src/app/custom-config.ts
@@ -1,5 +1,0 @@
-import { INGXLoggerConfig } from '../../../../src/public_api';
-
-export interface CustomConfig extends INGXLoggerConfig {
-    customProp: string;
-}