+++ projects/not-a-singleton/src/app/app-routing.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { LocalProviderComponent } from './local-provider/local-provider.component';
 import { MainComponent } from './main/main.component';
 
 const routes: Routes = [
@@ -11,6 +12,10 @@ const routes: Routes = [
   {
     path: 'not-a-singleton',
     loadChildren: (): any => import('projects/not-a-singleton/src/app/not-a-singleton/not-a-singleton.module').then(m => m.NotASingletonModule),
+  },
+  {
+    path: 'local-provider',
+    component: LocalProviderComponent,
   },
   {
     path: 'main',