+++ src/lib/config/iconfig-engine.ts
@@ -1,11 +1,6 @@
 import { NgxLoggerLevel } from './../types/logger-level.enum';
 import { INGXLoggerConfig } from './iconfig';
 
-/**
- * Injection token of logger config engine
- */
-export const TOKEN_LOGGER_CONFIG_ENGINE = 'TOKEN_LOGGER_CONFIG_ENGINE';
-
 export interface INGXLoggerConfigEngine {
 
   /** Get a readonly access to the level configured for the NGXLogger */