+++ testing/src/lib/custom-logger.service.mock.ts
@@ -1,5 +1,5 @@
-import { NGXLoggerMock } from './logger.service.mock';
 import { Injectable } from "@angular/core";
+import { NGXLoggerMock } from "./logger.service.mock";
 
 /**
  * CustomNGXLoggerServiceMock is a mock for CustomNGXLoggerService