+++ testing/src/lib/logger-testing.module.ts
@@ -1,18 +1,6 @@
 import { NgModule } from '@angular/core';
-import {
-  CustomNGXLoggerService,
-  LoggerModule,
-  NGXLogger,
-  NgxLoggerLevel,
-  TOKEN_LOGGER_CONFIG,
-  TOKEN_LOGGER_CONFIG_ENGINE,
-  TOKEN_LOGGER_MAPPER_SERVICE,
-  TOKEN_LOGGER_METADATA_SERVICE,
-  TOKEN_LOGGER_RULES_SERVICE,
-  TOKEN_LOGGER_SERVER_SERVICE,
-  TOKEN_LOGGER_WRITER_SERVICE
-} from 'ngx-logger';
-import { NGXLoggerConfigEngineMock } from './config-engine.mock';
+import { CustomNGXLoggerService, LoggerModule, NGXLogger, NgxLoggerLevel, TOKEN_LOGGER_CONFIG, TOKEN_LOGGER_CONFIG_ENGINE_FACTORY, TOKEN_LOGGER_MAPPER_SERVICE, TOKEN_LOGGER_METADATA_SERVICE, TOKEN_LOGGER_RULES_SERVICE, TOKEN_LOGGER_SERVER_SERVICE, TOKEN_LOGGER_WRITER_SERVICE } from 'ngx-logger';
+import { NGXLoggerConfigEngineFactoryMock } from './config-engine-factory.mock';
 import { CustomNGXLoggerServiceMock } from './custom-logger.service.mock';
 import { NGXLoggerMock } from './logger.service.mock';
 import { NGXLoggerMapperServiceMock } from './mapper.service.mock';
@@ -27,7 +15,7 @@ import { NGXLoggerWriterServiceMock } from './writer.service.mock';
   providers: [
     { provide: NGXLogger, useClass: NGXLoggerMock },
     { provide: TOKEN_LOGGER_CONFIG, useValue: { level: NgxLoggerLevel.ERROR } },
-    { provide: TOKEN_LOGGER_CONFIG_ENGINE, useClass: NGXLoggerConfigEngineMock },
+    { provide: TOKEN_LOGGER_CONFIG_ENGINE_FACTORY, useClass: NGXLoggerConfigEngineFactoryMock },
     { provide: TOKEN_LOGGER_METADATA_SERVICE, useClass: NGXLoggerMetadataServiceMock },
     { provide: TOKEN_LOGGER_RULES_SERVICE, useClass: NGXLoggerRulesServiceMock },
     { provide: TOKEN_LOGGER_MAPPER_SERVICE, useClass: NGXLoggerMapperServiceMock },