+++ testing/src/lib/logger.service.mock.ts
@@ -1,6 +1,6 @@
 import { HttpHeaders, HttpParams } from '@angular/common/http';
-import { INGXLoggerMonitor, INGXLoggerConfig, NgxLoggerLevel } from 'ngx-logger';
 import { Injectable } from "@angular/core";
+import { INGXLoggerConfig, INGXLoggerMonitor, NgxLoggerLevel } from 'ngx-logger';
 
 // todo bmtheo, there should be an interface or something to make sure this mock sticks to the real API
 @Injectable()