+++ testing/src/lib/metadata.service.mock.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { INGXLoggerConfig, INGXLoggerMetadata, INGXLoggerMetadataService, NgxLoggerLevel } from 'ngx-logger';
 
+
 @Injectable()
 export class NGXLoggerMetadataServiceMock implements INGXLoggerMetadataService {